replaces dpct namespace for the atomic operations
