--- tests/cuda2/dpct_output/includes/dw_cuda_soa.dp.cpp
@@ -1,1 +1,0 @@
-../dw_cuda_soa.dp.cpp
