+++ tests/cuda2/dpct_output/includes/dw_soa_sycl.cpp
@@ -1,0 +1,1 @@
+../dw_soa_sycl.cpp
